feat(bom): add depreciation_waste row as last BOM item
Compute total depreciation as sum of totalLengthM × depreciation_pct/100
across all length-bearing items (each product has its own rate). Append a
single depreciation_waste row at the end of the BOM with pieceLengthM=1
and qty in meters. Migration 0042 activates the product.

Fix: used totalLengthM (metres) not qty (pieces) — previous formula gave
3 m instead of the correct 12.3 m on a 307 m aluminium project at 4%.

+++ BE/mgp-service/app/schemas/bom.py
@@ -9,7 +9,7 @@ class BOMItemRead(BaseModel):
     section: str | None = None           # logical section for grouping (e.g. 'trapezoids', 'diagonals_external')
     pieceLengthM: float | None = None    # length of one piece (length-bearing rows only)
     totalLengthM: float | None = None    # qty × pieceLengthM (length-bearing rows only)
-    qty: int
+    qty: float   # int for most rows; float for depreciation_waste (meters)
     productId: str | None = None
     partNumber: str | None = None
     name: str | None = None
